Untangle some classes in iocore
- Isolate the following classes into their own files
  - NetEvent
  - NetHandler
  - NetVCOptions
  - PollCont
  - AcceptOptions
- Fix some includes other files to compile cleanly
