+++ iocore/net/PollCont.h
@@ -1,0 +1,43 @@
+/** @file
+
+  A brief file description
+
+  @section license License
+
+  Licensed to the Apache Software Foundation (ASF) under one
+  or more contributor license agreements.  See the NOTICE file
+  distributed with this work for additional information
+  regarding copyright ownership.  The ASF licenses this file
+  to you under the Apache License, Version 2.0 (the
+  "License"); you may not use this file except in compliance
+  with the License.  You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+ */
+
+#pragma once
+
+#include "I_Continuation.h"
+#include "I_Net.h"
+
+class NetHandler;
+struct PollDescriptor;
+
+struct PollCont : public Continuation {
+  NetHandler *net_handler;
+  PollDescriptor *pollDescriptor;
+  PollDescriptor *nextPollDescriptor;
+  int poll_timeout;
+
+  PollCont(Ptr<ProxyMutex> &m, int pt = net_config_poll_timeout);
+  PollCont(Ptr<ProxyMutex> &m, NetHandler *nh, int pt = net_config_poll_timeout);
+  ~PollCont() override;
+  int pollEvent(int, Event *);
+  void do_poll(ink_hrtime timeout);
+};
